+++ irida_sistr_results/SistrResultsException.py
@@ -1,5 +1,9 @@
+"""
+An Exception raised if there was an issue parsing SISTR results.
+"""
+
+
 class SistrResultsException(Exception):
-    """An Exception raised if there was an issue parsing SISTR results."""
 
     def __init__(self, msg):
         super().__init__(msg)