+++ irida_sistr_results/sistr_writer.py
@@ -11,13 +11,14 @@ from irida_sistr_results.version import __version__
 class SistrResultsWriter(object):
     """Abstract class resonsible for writing SISTR results to a table format"""
 
-    def __init__(self, irida_url, appname, username):
-        """
-        Construct a new SistrResultsWriter object corresponding to the passed irida_url
+    def __init__(self, irida_url, appname, username, sample_created_min_date=None):
+        """Construct a new SistrResultsWriter object corresponding to the passed irida_url
 
-        :param irida_url: The URL to the IRIDA instance, used to insert URLs into the table
-        :param appname: The application name.
-        :param username: The name of the user generating these results.
+        Args:
+            irida_url: The URL to the IRIDA instance, used to insert URLs into the table
+            appname: The application name.
+            username: The name of the user generating these results.
+            sample_created_min_date: The minimum date for including samples.
         """
         __metaclass__ = abc.ABCMeta
         self.irida_url = irida_url
@@ -25,6 +26,7 @@ class SistrResultsWriter(object):
         self.username = username
         self.row = 0
         self.end_of_project = False
+        self.sample_created_min_date = sample_created_min_date
 
     @abc.abstractmethod
     def _write_row(self, row):
@@ -56,10 +58,10 @@ class SistrResultsWriter(object):
         return self.row
 
     def set_row(self, row):
-        """
-        Sets the current row number
+        """Sets the current row number
 
-        :param row: The new row number
+        Args:
+            row: The new row number
         """
         self.row = row
 
@@ -103,23 +105,23 @@ class SistrResultsWriter(object):
         return timestamp.strftime('%Y-%m-%d %H:%M:%S')
 
     def _get_header_index(self, title):
-        """
-        Gets the particular index from the headers given the title.
+        """Gets the particular index from the headers given the title.
 
-        :param title: The title of the header column.
+        Args:
+            title: The title of the header column.
 
-        :return: The index into the header list.
+        Returns: The index into the header list.
         """
         return self._get_header_list().index(title)
 
     def _get_row_list(self, project, result):
-        """
-        Given the project number and result object, creates a list of relavent information to print per row.
+        """Given the project number and result object, creates a list of relavent information to print per row.
 
-        :param project: The current project identifier.
-        :param result:  The current SistrInfo result object.
+        Args:
+            project: The current project identifier.
+            result:  The current SistrInfo result object.
 
-        :return: A list of relevant information for the row
+        Return: A list of relevant information for the row
         """
         return [
             project,
@@ -153,14 +155,15 @@ class SistrResultsWriter(object):
         ]
 
     def _get_no_results_row_list(self, project, result):
-        """
-        Gets a list respresenting no/missing results for a sample.
+        """Gets a list respresenting no/missing results for a sample.
 
-        :param project: The current project identifier.
-        :param result:  The current SistrInfo result object.
+        Args:
+            project: The current project identifier.
+            result:  The current SistrInfo result object.
 
-        :return: A list of relevant information in the case of a no/missing result row.
+        Return: A list of relevant information in the case of a no/missing result row.
         """
+
         return [
             project,
             result.get_sample_name(),
@@ -193,12 +196,12 @@ class SistrResultsWriter(object):
         ]
 
     def write(self, sistr_results):
-        """
-        Writes out the results to an appropriate file with the appropriate format
+        """Writes out the results to an appropriate file with the appropriate format
 
-        :param sistr_results:  The SISTR results to write to a table.
-        :return: None
+        Args:
+            sistr_results:  The SISTR results to write to a table.
         """
+
         self.set_row(0)
         self._write_header(self._get_header_list())
         self.set_row(1)
@@ -226,16 +229,18 @@ class SistrResultsWriter(object):
 
         self._formatting()
 
+        sample_create_msg = ' using only samples created after ' + self.sample_created_min_date.strftime(
+            '%Y-%m-%d %H:%M:%S') if self.sample_created_min_date else ''
         self._write_row([
-            "Results generated from " + self.appname + " version=" + __version__ + " connecting to IRIDA=" + self.irida_url + " as user=" + self.username + " on date=" + datetime.now().strftime(
-                '%Y-%m-%d %H:%M:%S')])
+                            "Results generated from " + self.appname + " version=" + __version__ + " connecting to IRIDA=" + self.irida_url + " as user=" + self.username + " on date=" + datetime.now().strftime(
+                                '%Y-%m-%d %H:%M:%S') + sample_create_msg])
 
 
 class SistrCsvWriter(SistrResultsWriter):
     """An abstact writer used to create CSV/tab-delimited files"""
 
-    def __init__(self, irida_url, appname, username, out_file):
-        super(SistrCsvWriter, self).__init__(irida_url, appname, username)
+    def __init__(self, irida_url, appname, username, out_file, sample_created_min_date=None):
+        super(SistrCsvWriter, self).__init__(irida_url, appname, username, sample_created_min_date)
         out_file_h = open(out_file, 'w')
         self.writer = csv.writer(out_file_h, delimiter="\t", quotechar='"', quoting=csv.QUOTE_MINIMAL)
 
@@ -281,8 +286,8 @@ class SistrCsvWriter(SistrResultsWriter):
 class SistrCsvWriterShort(SistrCsvWriter):
     """Creates a shortened version of the results in a tab-delimited format"""
 
-    def __init__(self, irida_url, appname, username, out_file):
-        super(SistrCsvWriterShort, self).__init__(irida_url, appname, username, out_file)
+    def __init__(self, irida_url, appname, username, out_file, sample_created_min_date=None):
+        super(SistrCsvWriterShort, self).__init__(irida_url, appname, username, out_file, sample_created_min_date)
 
     def _get_header_list(self):
         return [
@@ -330,8 +335,8 @@ class SistrCsvWriterShort(SistrCsvWriter):
 class SistrExcelWriter(SistrResultsWriter):
     """A writer object for writing SISTR results to an excel spreadsheet"""
 
-    def __init__(self, irida_url, appname, username, out_file):
-        super(SistrExcelWriter, self).__init__(irida_url, appname, username)
+    def __init__(self, irida_url, appname, username, out_file, sample_created_min_date=None):
+        super(SistrExcelWriter, self).__init__(irida_url, appname, username, sample_created_min_date)
         self.workbook = xlsxwriter.Workbook(out_file, {'default_date_format': 'yyyy/mm/dd'})
         self.worksheet = self.workbook.add_worksheet()
         self.index_of_cgmlst_percent = self._get_header_list().index('cgMLST Percent Matching')
@@ -341,32 +346,32 @@ class SistrExcelWriter(SistrResultsWriter):
         self.percent_format = self.workbook.add_format({'num_format': '0.0%'})
 
     def _get_header_column_number(self, title):
-        """
-        Gets the particular column number from the headers given the title.
+        """Gets the particular column number from the headers given the title.
 
-        :param title: The title of the header column.
+        Args:
+            title: The title of the header column.
 
-        :return: The column number (starting with 1) from the header list.
+        Returns: The column number (starting with 1) from the header list.
         """
         return self._get_header_index(title) + 1
 
     def _get_header_column_letter(self, title):
-        """
-        Gets the particular column letter from the headers given the title.
+        """Gets the particular column letter from the headers given the title.
 
-        :param title: The title of the header column.
+        Args:
+            title: The title of the header column.
 
-        :return: The column letter (starting with A) from the header list.
+        Returns: The column letter (starting with A) from the header list.
         """
         return self._to_letter(self._get_header_index(title))
 
     def _range_stitle(self, title):
-        """
-        Gets the particular column letter range from the headers given a single title.
+        """Gets the particular column letter range from the headers given a single title.
 
-        :param title: The title of the header column.
+        Args:
+            title: The title of the header column.
 
-        :return: The column range (e.g., A:A) from the header list.
+        Returns: The column range (e.g., A:A) from the header list.
         """
         return self._range_title(title, title)
 
@@ -471,5 +476,4 @@ class SistrExcelWriter(SistrResultsWriter):
         self.worksheet.freeze_panes(1, 3)
 
     def close(self):
-        """Closes the workbook"""
         self.workbook.close()